Fix scope of lookup of redeclaration modifications
Checking if extends clauses are handled before redeclarations is already
done in ExtendsOrder.mo.

To make sure that redeclarations on symbols are looked up correctly, we
set the scope for lookup on any modifications before passing them on to
the next level.

To avoid infinite recursion in stringifying classes, we handle scope just
like parent and root (i.e. ignore it in the recursion).